# feat: small speed up to processing incoming dns records