--- a/src/zeroconf/_dns.py
+++ b/src/zeroconf/_dns.py
@@ -114,7 +114,7 @@ class DNSQuestion(DNSEntry):
 
     def __init__(self, name: str, type_: int, class_: int) -> None:
         super().__init__(name, type_, class_)
-        self._hash = hash((self.key, type_, self.class_))
+        self._hash = hash((self.key, type_, class_))
 
     def answered_by(self, rec: 'DNSRecord') -> bool:
         """Returns true if the question is answered by the record"""
@@ -251,7 +251,7 @@ class DNSAddress(DNSRecord):
         super().__init__(name, type_, class_, ttl, created)
         self.address = address
         self.scope_id = scope_id
-        self._hash = hash((self.key, type_, self.class_, address, scope_id))
+        self._hash = hash((self.key, type_, class_, address, scope_id))
 
     def write(self, out: 'DNSOutgoing') -> None:
         """Used in constructing an outgoing packet"""
@@ -296,7 +296,7 @@ class DNSHinfo(DNSRecord):
         super().__init__(name, type_, class_, ttl, created)
         self.cpu = cpu
         self.os = os
-        self._hash = hash((self.key, type_, self.class_, cpu, os))
+        self._hash = hash((self.key, type_, class_, cpu, os))
 
     def write(self, out: 'DNSOutgoing') -> None:
         """Used in constructing an outgoing packet"""
@@ -332,7 +332,7 @@ class DNSPointer(DNSRecord):
         super().__init__(name, type_, class_, ttl, created)
         self.alias = alias
         self.alias_key = alias.lower()
-        self._hash = hash((self.key, type_, self.class_, self.alias_key))
+        self._hash = hash((self.key, type_, class_, self.alias_key))
 
     @property
     def max_size_compressed(self) -> int:
@@ -376,7 +376,7 @@ class DNSText(DNSRecord):
     ) -> None:
         super().__init__(name, type_, class_, ttl, created)
         self.text = text
-        self._hash = hash((self.key, type_, self.class_, text))
+        self._hash = hash((self.key, type_, class_, text))
 
     def write(self, out: 'DNSOutgoing') -> None:
         """Used in constructing an outgoing packet"""
@@ -425,7 +425,7 @@ class DNSService(DNSRecord):
         self.port = port
         self.server = server
         self.server_key = server.lower()
-        self._hash = hash((self.key, type_, self.class_, priority, weight, port, self.server_key))
+        self._hash = hash((self.key, type_, class_, priority, weight, port, self.server_key))
 
     def write(self, out: 'DNSOutgoing') -> None:
         """Used in constructing an outgoing packet"""
@@ -476,7 +476,7 @@ class DNSNsec(DNSRecord):
         super().__init__(name, type_, class_, ttl, created)
         self.next_name = next_name
         self.rdtypes = sorted(rdtypes)
-        self._hash = hash((self.key, type_, self.class_, next_name, *self.rdtypes))
+        self._hash = hash((self.key, type_, class_, next_name, *self.rdtypes))
 
     def write(self, out: 'DNSOutgoing') -> None:
         """Used in constructing an outgoing packet."""