--- a/src/zeroconf/_history.pxd
+++ b/src/zeroconf/_history.pxd
@@ -12,7 +12,7 @@ cdef class QuestionHistory:
     cpdef add_question_at_time(self, DNSQuestion question, float now, cython.set known_answers)
 
     @cython.locals(than=cython.double, previous_question=cython.tuple, previous_known_answers=cython.set)
-    cpdef suppresses(self, DNSQuestion question, cython.double now, cython.set known_answers)
+    cpdef bint suppresses(self, DNSQuestion question, cython.double now, cython.set known_answers)
 
     @cython.locals(than=cython.double, now_known_answers=cython.tuple)
     cpdef async_expire(self, cython.double now)