--- a/src/zeroconf/_protocol/incoming.pxd
+++ b/src/zeroconf/_protocol/incoming.pxd
@@ -65,6 +65,7 @@ cdef class DNSIncoming:
     cdef cython.float _now_float
     cdef public object scope_id
     cdef public object source
+    cdef public bint _has_qu_question
 
     @cython.locals(
         question=DNSQuestion