… src/zeroconf/_services/browser.pxd
… src/zeroconf/_services/browser.py
… src/zeroconf/_services/info.pxd
… tests/test_protocol.py