roughly a ~7% speed up for incoming
before: Parsing 100000 incoming messages took 2.558975583058782 seconds
after: Parsing 100000 incoming messages took 2.3616396670695394 seconds

roughly a ~2% speed up for outgoing (mostly a side effect)
before: Construction 100000 outgoing messages took 1.9390459168935195 seconds
after: Construction 100000 outgoing messages took 1.8946991249686107 seconds